StudyClan | Login and register

+++ base/views.py
@@ -1,5 +1,11 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 # Create your views here.
@@ -9,12 +15,57 @@ from .forms import RoomForm
 #   { 'id': 2, 'name': 'Let\'s learn React' },
 #   { 'id': 3, 'name': 'React Native Developers Assemble' },
 # ]
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        messages.info(request, 'You are already logged in')
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
 
+        try: 
+            user = User.objects.get(username=username)
+        except User.DoesNotExist:
+            messages.error(request, 'User does not exist')
+            return redirect('login')
+        user = authenticate(request, username=username, password=password)    
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password is incorrect')
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.user.is_authenticated:
+        messages.info(request, 'You are already logged in')
+        return redirect('home')
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error has occured during registration')
+            return redirect('register')
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context)
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q) | Q(host__username__icontains=q))
     room_count = rooms.count()
-    print(rooms)
     topics = Topic.objects.all()
     context = { 'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
@@ -32,6 +83,7 @@ def room(request, roomid):
     else:
         return render(request, 'base/room.html', context)
     
+@login_required(login_url='login')    
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -43,9 +95,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')  
 def updateRoom(request, roomid):
     room = Room.objects.get(id=roomid)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this room')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -56,8 +112,12 @@ def updateRoom(request, roomid):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')  
 def deleteRoom(request, roomid):
     room = Room.objects.get(id=roomid)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to delete this room')
     
     if request.method == 'POST':
         room.delete()
